+++ R/learner_randomForestSRC_classif_imbalanced_rfsrc.R
@@ -94,7 +94,8 @@ LearnerClassifImbalancedRandomForestSRC = R6Class("LearnerClassifImbalancedRando
         ptn.count = p_int(default = 0L, lower = 0L, tags = "predict"),
         cores = p_int(default = 1L, lower = 1L, tags = c("train", "predict", "threads")),
         save.memory = p_lgl(default = FALSE, tags = "train"),
-        perf.type = p_fct(levels = c("gmean", "misclass", "brier", "none"), tags = "train") # nolint
+        perf.type = p_fct(levels = c("gmean", "misclass", "brier", "none"), tags = "train"), # nolint
+        case.depth = p_lgl(default = FALSE, tags = "predict")
       )
 
       super$initialize(