+++ tests/testthat/test_catboost_classif_catboost.R
@@ -16,3 +16,14 @@ test_that("#353", {
   pred = rr$prediction() 
   assert_true(all(abs(rowSums(pred$prob) - 1) < 0.001))
 })
+
+test_that("early stopping works", {
+  learner = lrn("classif.catboost", iterations = 1000, eval_metric = "Logloss", early_stopping_rounds = 10)
+  learner$validate = 0.2
+  task = tsk("pima")
+
+  learner$train(task)
+
+  expect_list(learner$internal_tuned_values)
+  expect_number(learner$internal_tuned_values$iterations)
+})